test(jans-cedarling): move tests to a separate file and enhance input clarity
- reorganized tests into a dedicated file for better structure
- improved readability of policy and schema inputs in the tests
