+++ jans-cedarling/cedarling/src/common/policy_store.rs
@@ -5,6 +5,9 @@
  * Copyright (c) 2024, Gluu, Inc.
  */
 
+#[cfg(test)]
+mod test;
+
 use super::cedar_schema::CedarSchema;
 use base64::prelude::*;
 use cedar_policy::PolicyId;
@@ -288,141 +291,4 @@ where
         })?;
 
     Ok(policy)
-}
-
-#[cfg(test)]
-mod test {
-    use serde_json::json;
-
-    use super::ParsePolicySetMessage;
-    use super::PolicyStore;
-    use crate::common::policy_store::parse_and_check_token_metadata;
-    use crate::common::policy_store::parse_cedar_version;
-
-    /// Tests successful deserialization of a valid policy store JSON.
-    #[test]
-    fn test_policy_store_deserialization_success() {
-        static POLICY_STORE_RAW: &str = include_str!("../../../test_files/policy-store_ok.json");
-        assert!(serde_json::from_str::<PolicyStore>(POLICY_STORE_RAW).is_ok());
-    }
-
-    /// Tests for base64 decoding error in the policy store.
-    #[test]
-    fn test_base64_decoding_error_in_policy_store() {
-        static POLICY_STORE_RAW: &str =
-            include_str!("../../../test_files/policy-store_policy_err_base64.json");
-
-        let policy_result = serde_json::from_str::<PolicyStore>(POLICY_STORE_RAW);
-        assert!(policy_result
-            .unwrap_err()
-            .to_string()
-            .contains(&ParsePolicySetMessage::Base64.to_string()));
-    }
-
-    /// Tests for parsing error due to broken UTF-8 in the policy store.
-    #[test]
-    fn test_policy_parsing_error_in_policy_store() {
-        static POLICY_STORE_RAW: &str =
-            include_str!("../../../test_files/policy-store_policy_err_broken_utf8.json");
-
-        let policy_result = serde_json::from_str::<PolicyStore>(POLICY_STORE_RAW);
-        assert!(policy_result
-            .unwrap_err()
-            .to_string()
-            .contains(&ParsePolicySetMessage::String.to_string()));
-    }
-
-    /// Tests for broken policy parsing error in the policy store.
-    #[test]
-    fn test_broken_policy_parsing_error_in_policy_store() {
-        static POLICY_STORE_RAW: &str =
-            include_str!("../../../test_files/policy-store_policy_err_broken_policy.json");
-
-        let policy_result = serde_json::from_str::<PolicyStore>(POLICY_STORE_RAW);
-        let err_msg = policy_result.unwrap_err().to_string();
-
-        // TODO: this isn't really a human readable format so idk why the error message is
-        // like this. Look into this in the future.
-        assert_eq!(err_msg, "Errors encountered while parsing policies: [Error(\"unable to decode policy with id: 840da5d85403f35ea76519ed1a18a33989f855bf1cf8, error: unable to decode policy_content from human readable format: unexpected token `)`\", line: 0, column: 0)] at line 9 column 5")
-    }
-
-    /// Tests that a valid version string is accepted.
-    #[test]
-    fn test_valid_version() {
-        let valid_version = "1.2.3".to_string();
-        assert!(parse_cedar_version(serde_json::Value::String(valid_version)).is_ok());
-    }
-
-    /// Tests that a valid version string with 'v' prefix is accepted.
-    #[test]
-    fn test_valid_version_with_v() {
-        let valid_version_with_v = "v1.2.3".to_string();
-        assert!(parse_cedar_version(serde_json::Value::String(valid_version_with_v)).is_ok());
-    }
-
-    /// Tests that an invalid version format is rejected.
-    #[test]
-    fn test_invalid_version_format() {
-        let invalid_version = "1.2".to_string();
-        assert!(parse_cedar_version(serde_json::Value::String(invalid_version)).is_err());
-    }
-
-    /// Tests that an invalid version part (non-numeric) is rejected.
-    #[test]
-    fn test_invalid_version_part() {
-        let invalid_version = "1.two.3".to_string();
-        assert!(parse_cedar_version(serde_json::Value::String(invalid_version)).is_err());
-    }
-
-    /// Tests that an invalid version format with 'v' prefix is rejected.
-    #[test]
-    fn test_invalid_version_format_with_v() {
-        let invalid_version_with_v = "v1.2".to_string();
-        assert!(parse_cedar_version(serde_json::Value::String(invalid_version_with_v)).is_err());
-    }
-
-    /// Tests that an error is returned for multiple role mappings in token metadata.
-    #[test]
-    fn test_invalid_multiple_role_mappings_in_token_metadata() {
-        let invalid_token_metadata = json!([
-            { "type": "Access", "person_id": "aud" },
-            { "type": "Id", "person_id": "sub", "role_mapping": "role" },
-            { "type": "userinfo", "person_id": "email", "role_mapping": "role" }
-        ]);
-
-        assert!(
-            parse_and_check_token_metadata(invalid_token_metadata).is_err(),
-            "expected an error for multiple role mappings"
-        );
-    }
-
-    /// Tests successful parsing of role mappings in token metadata.
-    #[test]
-    fn test_successful_parsing_of_role_mappings() {
-        let valid_token_metadata = json!([
-            { "type": "Access_token", "person_id": "aud" },
-            { "type": "Id_token", "person_id": "sub", "role_mapping": "role" },
-            { "type": "userinfo_token", "person_id": "email" }
-        ]);
-
-        assert!(
-            parse_and_check_token_metadata(valid_token_metadata).is_ok(),
-            "expected successful parsing of role mappings"
-        );
-    }
-
-    /// Tests unsuccessful parsing of role mappings in token metadata.
-    #[test]
-    fn test_error_on_invalid_token_type() {
-        let invalid_token_metadata = json!([
-            { "type": "Access_token", "person_id": "aud" },
-            { "type": "unknown_token", "person_id": "sub", "role_mapping": "role" },
-            { "type": "userinfo_token", "person_id": "email" }
-        ]);
-
-        assert!(
-            parse_and_check_token_metadata(invalid_token_metadata).is_err(),
-            "expected unsuccessful parsing of role mappings"
-        );
-    }
 }
